# increased performance of frameAndSend