--- a/bench/util.js
+++ b/bench/util.js
@@ -10,7 +10,7 @@
  * Returns a Buffer from a "ff 00 ff"-type hex string.
  */
 
-getBufferFromHexString = function(byteStr) {
+global.getBufferFromHexString = function(byteStr) {
   var bytes = byteStr.split(' ');
   var buf = new Buffer(bytes.length);
   for (var i = 0; i < bytes.length; ++i) {
@@ -23,7 +23,7 @@ getBufferFromHexString = function(byteStr) {
  * Returns a hex string from a Buffer.
  */
 
-getHexStringFromBuffer = function(data) {
+global.getHexStringFromBuffer = function(data) {
   var s = '';
   for (var i = 0; i < data.length; ++i) {
     s += padl(data[i].toString(16), 2, '0') + ' ';
@@ -35,7 +35,7 @@ getHexStringFromBuffer = function(data) {
  * Splits a buffer in two parts.
  */
 
-splitBuffer = function(buffer) {
+global.splitBuffer = function(buffer) {
   var b1 = new Buffer(Math.ceil(buffer.length / 2));
   buffer.copy(b1, 0, 0, b1.length);
   var b2 = new Buffer(Math.floor(buffer.length / 2));
@@ -47,7 +47,7 @@ splitBuffer = function(buffer) {
  * Performs hybi07+ type masking on a hex string or buffer.
  */
 
-mask = function(buf, maskString) {
+global.mask = function(buf, maskString) {
   if (typeof buf == 'string') buf = new Buffer(buf);
   var mask = getBufferFromHexString(maskString || '34 83 a8 68');
   for (var i = 0; i < buf.length; ++i) {
@@ -60,7 +60,7 @@ mask = function(buf, maskString) {
  * Returns a hex string representing the length of a message
  */
 
-getHybiLengthAsHexString = function(len, masked) {
+global.getHybiLengthAsHexString = function(len, masked) {
   if (len < 126) {
     var buf = new Buffer(1);
     buf[0] = (masked ? 0x80 : 0) | len;
@@ -82,7 +82,7 @@ getHybiLengthAsHexString = function(len, masked) {
  * Unpacks a Buffer into a number.
  */
 
-unpack = function(buffer) {
+global.unpack = function(buffer) {
   var n = 0;
   for (var i = 0; i < buffer.length; ++i) {
     n = (i == 0) ? buffer[i] : (n * 256) + buffer[i];
@@ -94,7 +94,7 @@ unpack = function(buffer) {
  * Returns a hex string, representing a specific byte count 'length', from a number.
  */
 
-pack = function(length, number) {
+global.pack = function(length, number) {
   return padl(number.toString(16), length, '0').replace(/([0-9a-f][0-9a-f])/gi, '$1 ').trim();
 }
 
@@ -102,6 +102,6 @@ pack = function(length, number) {
  * Left pads the string 's' to a total length of 'n' with char 'c'.
  */
 
-padl = function(s, n, c) {
+global.padl = function(s, n, c) {
   return new Array(1 + n - s.length).join(c) + s;
 }