--- a/lib/Sender.js
+++ b/lib/Sender.js
@@ -142,8 +142,8 @@ class Sender extends EventEmitter {
 
     if (!Buffer.isBuffer(data)) {
       canModifyData = true;
-      if (data && (typeof data.byteLength !== 'undefined' || typeof data.buffer !== 'undefined')) {
-        data = getArrayBuffer(data);
+      if (data && (data.buffer || data) instanceof ArrayBuffer) {
+        data = getBufferFromNative(data);
       } else {
         //
         // If people want to send a number, this would allocate the number in
@@ -268,7 +268,7 @@ class Sender extends EventEmitter {
   applyExtensions(data, fin, compress, callback) {
     if (compress && data) {
       if ((data.buffer || data) instanceof ArrayBuffer) {
-        data = getArrayBuffer(data);
+        data = getBufferFromNative(data);
       }
       this.extensions[PerMessageDeflate.extensionName].compress(data, fin, callback);
     } else {
@@ -279,16 +279,12 @@ class Sender extends EventEmitter {
 
 module.exports = Sender;
 
-function getArrayBuffer(data) {
+function getBufferFromNative(data) {
   // data is either an ArrayBuffer or ArrayBufferView.
-  var array = new Uint8Array(data.buffer || data),
-    l = data.byteLength || data.length,
-    o = data.byteOffset || 0,
-    buffer = new Buffer(l);
-  for (var i = 0; i < l; ++i) {
-    buffer[i] = array[o + i];
+  if (data.buffer) {
+    data = data.buffer.slice(data.byteOffset, data.byteOffset + data.byteLength);
   }
-  return buffer;
+  return new Buffer(data);
 }
 
 function getRandomMask() {