--- a/lib/Sender.js
+++ b/lib/Sender.js
@@ -281,10 +281,9 @@ module.exports = Sender;
 
 function getBufferFromNative(data) {
   // data is either an ArrayBuffer or ArrayBufferView.
-  if (data.buffer) {
-    data = data.buffer.slice(data.byteOffset, data.byteOffset + data.byteLength);
-  }
-  return new Buffer(data);
+  return !data.buffer
+    ? new Buffer(data)
+    : new Buffer(data.buffer).slice(data.byteOffset, data.byteOffset + data.byteLength)
 }
 
 function getRandomMask() {