--- a/lib/Sender.js
+++ b/lib/Sender.js
@@ -127,24 +127,17 @@ class Sender extends EventEmitter {
     var canModifyData = false;
 
     if (!data) {
-
-      try {
-        var buff = [opcode | (finalFragment ? 0x80 : 0), 0 | (maskData ? 0x80 : 0)]
-          .concat(maskData ? [0, 0, 0, 0] : []);
-        this._socket.write(new Buffer(buff), 'binary', cb);
-      }
-      catch (e) {
-        if (typeof cb == 'function') cb(e);
-        else this.emit('error', e);
-      }
+      var buff = [opcode | (finalFragment ? 0x80 : 0), 0 | (maskData ? 0x80 : 0)]
+        .concat(maskData ? [0, 0, 0, 0] : []);
+      sendFramedData.call(this, new Buffer(buff), null, cb);
       return;
     }
 
     if (!Buffer.isBuffer(data)) {
-      canModifyData = true;
       if (data && (data.buffer || data) instanceof ArrayBuffer) {
         data = getBufferFromNative(data);
       } else {
+        canModifyData = true;
         //
         // If people want to send a number, this would allocate the number in
         // bytes as memory size instead of storing the number as buffer value. So
@@ -194,49 +187,17 @@ class Sender extends EventEmitter {
       outputBuffer[dataOffset - 1] = mask[3];
       if (mergeBuffers) {
         bufferUtil.mask(data, mask, outputBuffer, dataOffset, dataLength);
-        try {
-          this._socket.write(outputBuffer, 'binary', cb);
-        }
-        catch (e) {
-          if (typeof cb == 'function') cb(e);
-          else this.emit('error', e);
-        }
-      }
-      else {
+      } else {
         bufferUtil.mask(data, mask, data, 0, dataLength);
-        try {
-          this._socket.write(outputBuffer, 'binary');
-          this._socket.write(data, 'binary', cb);
-        }
-        catch (e) {
-          if (typeof cb == 'function') cb(e);
-          else this.emit('error', e);
-        }
       }
     }
     else {
       outputBuffer[1] = secondByte;
       if (mergeBuffers) {
         data.copy(outputBuffer, dataOffset);
-        try {
-          this._socket.write(outputBuffer, 'binary', cb);
-        }
-        catch (e) {
-          if (typeof cb == 'function') cb(e);
-          else this.emit('error', e);
-        }
-      }
-      else {
-        try {
-          this._socket.write(outputBuffer, 'binary');
-          this._socket.write(data, 'binary', cb);
-        }
-        catch (e) {
-          if (typeof cb == 'function') cb(e);
-          else this.emit('error', e);
-        }
       }
     }
+    sendFramedData.call(this, outputBuffer, mergeBuffers ? null : data, cb);
   }
 
   /**
@@ -293,4 +254,19 @@ function getRandomMask() {
     ~~(Math.random() * 255),
     ~~(Math.random() * 255)
   ]);
+}
+
+function sendFramedData(outputBuffer, data, cb) {
+  try {
+    if (data) {
+      this._socket.write(outputBuffer, 'binary');
+      this._socket.write(data, 'binary', cb);
+    } else {
+      this._socket.write(outputBuffer, 'binary', cb);
+    }
+  }
+  catch (e) {
+    if (typeof cb == 'function') cb(e);
+    else this.emit('error', e);
+  }
 }