--- a/lib/WebSocket.js
+++ b/lib/WebSocket.js
@@ -436,7 +436,7 @@ WebSocket.prototype.addEventListener = function(method, listener) {
 
   function onMessage (data, flags) {
     if (flags.binary && this.binaryType === 'arraybuffer')
-      data = new Uint8Array(data).buffer;
+      data = data.buffer.slice(data.byteOffset, data.byteOffset + data.length);
     listener.call(target, new MessageEvent(data, !!flags.binary, target));
   }