The V8 engine optimizes code on a per-function basis. But it can't optimize functions that have a `try {} catch {}` block in them. Because of this, `frameAndSend()` was not getting optimized. And I don't know about you, but I'd say that's a pretty important function for optimization.

To fix this, I moved the `try {} catch {}` block into a separate function.

Also, when converting from ArrayBuffers/TypedArrays to Buffers, the data was being copied in memory, which has a speed of _O(n)_. With this pull request, converting between the two will no longer copy the memory, which is much faster; _O(1)_.

Also, I renamed the function `getArrayBuffer ` to `getBufferFromNative ` to better describe what it actually does.