@JoshuaWise 2016-08-20
Disclaimer:
If there's an actual reason for the memory to be copied, please inform me.