@JoshuaWise 2016-08-20
I ran *sender.benchmark.js*...

Before:
```
  Sender frameAndSend, unmasked (200 kB)
  3348536.80 ops/sec, 261996 times executed, benchmark took 5.1 sec.

  Sender frameAndSend, masked (200 kB)
  7371.89 ops/sec, 976 times executed, benchmark took 5.085 sec.
```

After:
```
  Sender frameAndSend, unmasked (200 kB)
  4045681.98 ops/sec, 314701 times executed, benchmark took 5.031 sec.

  Sender frameAndSend, masked (200 kB)
  7438.43 ops/sec, 976 times executed, benchmark took 5.037 sec.
```